all working types, need tests and panic handling

+++ helpers.go
@@ -6,7 +6,7 @@ import (
 
 // percentDo takes a percent (float32 between 0 and 1)
 // and randomly returns true that percent of the time
-func percentDo(p float32) bool {
+func percentDo(p float64) bool {
 	var proceed bool
 
 	// bias false if p < 0.0, p > 1.0
@@ -15,7 +15,7 @@ func percentDo(p float32) bool {
 	}
 
 	// 0.0 <= r < 1.0
-	r := rand.Float32()
+	r := rand.Float64()
 	if r < p {
 		return true
 	}
